+++ modules/canvaskit/npm_build/types/index.d.ts
@@ -3296,6 +3296,12 @@ export interface ColorFilterFactory {
      * Makes a color filter that converts between sRGB colors and linear colors.
      */
     MakeSRGBToLinearGamma(): ColorFilter;
+
+    /**
+     * Makes a color filter that multiplies the luma of its input into the alpha channel,
+     * and sets the red, green, and blue channels to zero.
+     */
+    MakeLuma(): ColorFilter;
 }
 
 export interface ContourMeasureIterConstructor {